refactor(edihistory): lift edih_x12_file to OpenEMR\Billing\EdiHistory\X12File
Move the 1610-line edih_x12_file class body from
library/edihistory/edih_x12file_class.php into PSR-4 location
src/Billing/EdiHistory/X12File.php under the
OpenEMR\Billing\EdiHistory namespace. The legacy file becomes a
class_alias() shim so existing procedural callers across
library/edihistory/* keep resolving the unqualified `edih_x12_file`
symbol unchanged.

Behavior-preserving:
- No type tightening on properties, parameters, or return types
- Global \text() and \finfo() now fully qualified for namespace
  resolution (no semantic change)
- New file declares strict_types=1 (no behavior impact since the
  class never relied on coercion at boundaries)

PHPStan support:
- New .phpstan/phpstan_legacy_aliases.php bootstrap mirrors the
  runtime class_alias so static analysis can resolve `edih_x12_file`
  references in legacy callers (PHPStan does not follow runtime
  class_alias calls)
- Baseline regenerated; fatal-cap counts unchanged
  (class.notFound 229, method.notFound 141, variable.undefined 3064)

Tests:
- New isolated test suite at
  tests/Tests/Isolated/Billing/EdiHistory/X12FileIsolatedTest.php
  exercises the parser against a synthetic minimal X12 270 fixture
  built in setUp(); verifies envelope detection, delimiter parsing,
  segment population, and text-retention behavior without requiring
  the legacy bootstrap or any PHI fixtures

Refs openemr/openemr#11792.
